tmux: Add a caveat on need for `kill-server`
Each time tmux is upgraded, it is generally a good idea to kill any
existing sessions to avoid incompatibilities between new client and
sessions from an server. Leading to a crash such as in issue #97955

Add a note to the caveats of the tmux formula noting that users may need
to run 'tmux kill-server' after upgrading the tmux formula.