rabbitmq: explain how to enable all feature flags
in a post-installation info message.

Without this hint, many Homebrew users will end up
with a node that only has a subset of feature flags
enabled. Such installations cannot be upgraded
to RabbitMQ 3.12.x and future 3.13.x.

As a result, the formula installs successfully but
the node refuses to start.

Also link to [1][2], which explain how feature
flags are used in RabbitMQ and why enabling them
after installation or upgrade is important.

Future RabbitMQ versions will display a management UI
warning when at least one stable feature flag is
not enabled.

1. https://rabbitmq.com/feature-flags.html
2. https://rabbitmq.com/upgrade.html