homebrew-core disable triage staleness updates
This commit changes the `triage-issues` workflow to no longer
automatically remove the `stale` tag when updated. Instead, project
managers with proper permissions will need to remove the label manually
whenever that is deemed necessary.

This automation will reduce the number of volunteer contact points
required to handle tickets that have not received attention from our
more knowledgeable and code-contributing volunteers and ensure their
time isn't wasted helping aspiring volunteers to figure out how to
submit valid PRs to resolve tickets. This will keep the number of
unresolved tickets low, which is of course the most important metric for
an OSS community-driven project like ours.

Thanks to the project volunteer who, although unable to provide help for
an obviously broken and easily fixed package[0] (#153199), was able to
spare time to close my "really long stale" issue that "nobody has any
idea how to fix". He helpfully pointed out that despite my included
patch, the issue should have been "closed ages ago" because:

> There are no additional support channels. If no volunteer is
> interested in your issue then no volunteer time will be spend[sic] on
> it, it's that simple.

After meditating on this, I realized the wisdom of his words as they
pointed out a fallacy: Indeed, time IS spent on every single issue that
remains open, when the issue opener cares enough to keep bumping it.
It's the time spent closing the issue, like he had done! So clever, and
such a hackerly way to point that out to me.

In the same hacker spirit, I present this PR to fix the underlying
issue. Alas I have also not tested this patch, this time not because of
inscrutable build errors (I mean it's a missing system library, right?),
but rather because I've decided to try to learn Nix again and I have to
figure out how to get a flake to drop a service run script in the right
directory, and it's 2024 and I'm in hell.

init save me.

~eblume

PS I am looking for BE/DevOps employment for fully remote positions.

[0]: https://github.com/Homebrew/homebrew-core/issues/153199