stellar-core: update to version 20.3.0
Update the formula to install stellar-core version 20.3.0.

Requires adding the coreutils dependency because the install scripts use the
`sha256sum` command.

Requires adding the rust dependency because the application is since version 20
a combinarion of C++ and Rust.

The configure options were reduced and now relies on std_configure_args.

The upgrades test category was removed because the test fails due to a string
formatting comparison error on one target (macOS 13 x64). The test categories
that have been selected to run after install appear to be arbitrary. There are
120 test categories in total (according to stellar-core test -t). I don't see
much reason to select the above four over others that are available. It's
possible this small subset were selected to get some signal that the binary
works. If that's the case, I don't think it's a concern to remove upgrades.

Co-authored-by: Sean Molenaar <SMillerDev@users.noreply.github.com>