Update libxc.rb
A bigger software will use libxc in the future, but with this functions. The development of this software is on a macos-system. 
As i see inside the ChangeLog.md

"- Added the possibility of selectively disabling the compilation of vxc, fxc, kxc, and lxc to both Autotools and CMake build systems. By default kxc and lxc are not compiled. This allows for faster compilation times."

this parameter is inside just for reducing the compiling-time.