tcpflow: update stable, livecheck
The downloads.digitalcorpora.org URLs in the `tcpflow` formula now
redirect to corp.digitalcorpora.org. This updates the `stable` and
`livecheck` block URLs to avoid the redirection.

The `stable` URL subsequently redirects to
digitalcorpora.s3.amazonaws.com but there may be something to be said
for keeping the digitalcorpora.org URL, as it would continue to work
if upstream switches to a different file host and points the
redirection somewhere else.