aws-sam-cli: add livecheck
By default, livecheck uses the `Pypi` strategy for `aws-sam-cli` but
the generated strategy regex fails to match the tarball URL because
the filename prefix of the current tarball uses underscores
(`aws_sam_cli`) instead of hyphens (`aws-sam-cli`). This adds a
`livecheck` block that allows for either option in the filename
prefix.