cloudflare-cli4 2.19.4 (new formula)
This commit adds a new formula for Cloudflare's official CLI tool for the version 4 of the API.
The tool is fairly complex and can be used to interact with the API is various ways, some examples can be found in the README[0].

The project repository is at https://github.com/cloudflare/python-cloudflare-cli4.

`cli4`, the command line tool, was previously packaged as part of the `cloudflare` PyPi package but will now be packaged separately.

The homebrew package should be equivalent to the version published at PyPi[1] under the name cloudflare-cli4.

[0]: https://github.com/cloudflare/python-cloudflare-cli4/blob/main/README.md
[1]: https://pypi.org/project/cloudflare-cli4/