mariadb: fix binding to 127.0.0.1 by default
Currently the configuration to bind the server to 127.0.0.1 is not applied correctly.

This is because following the installation process there will be already a `my.cnf` file before installing the `my.cnf` file containing the bind-address configuration. Thus keeping the first `my.cnf` and installing to `my.cnf.default` instead.

To fix this issue, install to bind-address configuration file to `my.cnf.d/` instead.