semgrep: restrict pip version <24
Our nightly builds that use brew to build semgrep from HEAD have
started failing recently. We've noticed that it's related to the new
version of pip (v24) being picked up in many environments.

While we figure out how to make semgrep build properly with the newest
pip, we'd like to unblock these workflows, which include the releases
too.

This PR pins the pip version to <24.