semgrep: skip opam dependency checking for brew in linux
Brew in linux doesn't seem compatible with opam. I believe opam uses
the system's default package manager to check dependencies.

So in this case, brew on linux installs dependencies, but opam thinks
that they're not installed and gives an error.

Worked around this by only checking dependencies on macos. Not ideal
for linux, but if it's missing a dependency, the build should fail,
and if the build doesn't fail, then brew test should fail.