mackup: deprecate
`mackup` works by moving config files to an
off-disk backup location (usually a cloud storage
provider like iCloud or Dropbox, etc.) and then
replacing the original file with a symlink to the
backed up version. The intended use is to make
sure program configuration settings and
preferences can be restored if you need to replace
computers, factory reset, etc.

As documented
[here](https://github.com/lra/mackup/issues/2035),
[here](https://github.com/lra/mackup/issues/1924),
[here](https://github.com/lra/mackup/discussions/1969#discussioncomment-8485995),
`mackup backup` no longer works on MacOS Sonoma
(likely due to [security
changes](https://github.com/lra/mackup/issues/1924#issuecomment-2026186178)
in the app sandbox in Sonoma) and running either
the backup or restore commands on MacOS Sonoma
will result in the loss of all preference files.
Users who backup to a storage provider that has
versioning or backup snapshots may be able to
restore preferences to an older version. For users
backing up to other storage providers that do not
have this feature, such as iCloud, the command
results in total data loss. This issue has been
known to the developer for over a year but there
is no fix, no mention of the issue in the project
README, no warning when running the command, and
no indication in the program output that the
backup process may not have completed properly -
neither in the dry-run mode nor in live mode.