kertish-dfs: remove version override
The `kertish-dfs` formula was originally introduced by the author
with the version format used in tags (e.g., 21.2.0066). However, the
formula was updated to use a version format with a suffix
(22.2.0147-532592) and the suffix doesn't appear to be available
anywhere upstream (e.g., tags, release notes, code, etc.). As a
result, livecheck always treats the formula version as newer than the
upstream version, as the upstream tags don't include the suffix.

If the suffix is an important part of the version, then upstream
should either add it to the tag version or include it somewhere in
the release information. We can't support a version format where all
the version information isn't available, so this removes the suffix
from the `version` to bring it in line with the format in tags.