libtirpc: make formula work on macos
this changes the formula to not compile only on linux. on macos, the bundled
kerberos is used. because the necessary changes for this to work are not
yet upstream, configure must be regenerated, so aclocal, autoconf and automake
will need to be run;

this includes an inline copy of patches which were submitted to libtirpc-devel
list for inclusion (upstream does not use github so i had to do this rather
than PR)