fbthrift: make `mvfst` and `boost` build- and test-only
This has no linkage with `mvfst` and `boost` on macOS, so we can likely
make them build-only.

There is linkage to boost on Linux so we keep it as a runtime dep there.

Also, let's use `python3` from macOS to build, since this doesn't need a
new version of Python.