workflows/tests: set HOMEBREW_LIVECHECK_AUTOBUMP
We recently modified the behavior of `brew livecheck` to skip
packages in the tap's autobump list, so this change is causing
autobumped formulae to fail the `brew livecheck` part of CI testing.
This adds the `HOMEBREW_LIVECHECK_AUTOBUMP` ENV variable, which
will ensure that `brew livecheck` doesn't skip autobumped formulae
(without manually passing the `--autobump` flag).