tgenv 1.2.1 (switch to maintained fork)
tgenv (update repo)

The current tgenv (https://github.com/cunymatthieu/tgenv) is not maintained
and hasn't had a release since 2019. It was forked to
https://github.com/tgenv/tgenv in 2023 and is actively worked on.

Update formula to this new repo

tgenv: fix test

Existing test passes on Linux but fails on at least one version of MacOs.
Fix it so that it checks it can retrieve the oldest and newest versions,
rather than an arbitrary "middle" one