BV-BRC-CLI 1.043 (new formula)
The BV-BRC-CLI module is a distribution for the command-line interface to the
Bacterial and Viral Bioinformatics Resource Center website and analysis
tools. The site is funded by the US National Institute of Health's Institute
for Allergy and Infectious Diseases (NIH NIAID).

This submission passes brew --audit; however it fails the notability test in
brew audit --new. This is because the github module BV-BRC/BV-BRC-CLI
is a wrapper module that contains release engineering data for the
creation of production releases for a collection of other modules,
largely hosted at https://github.com/BV-BRC. These modules have been
in use for several years and have more activity, though they have
not had forks/watchers/stars of the volume requested.

Our target audience is not necessarily github-savvy, which is why
we wish to be included in homebrew. Our current release is via a
Mac app wrapper, which is becoming more unwieldy to manage due to
increasing limitations on app installation under Sequoia. We also
release via debain and redhat/fedora packages, though those also
have limitations for users who may not be allowed to install
such packages.