gnupg.rb: replace TLS dependency with in-house solution
the dependency on gnutls draws in a lot of secondary dependencies which in total amount to almost half of the time needed for a full build. the GnuPG project developed ntbTLS as a light-weight substitution that also does the job without depending on any additional libraries (at least none that aren't also needed by GnuPG itself as well).