guichan 0.8.3
This re-introduces the previously deprecated, disabled and removed
formula for Guichan and updates it to the new official location and
recent 0.8.3 release.

It also removes the dependencies on sdl_image. The Guichan core library
is generic and still useful to games not based on SDL 1. The test was
replaced with an example using Guichan without any of its extension
libraries.

This reverts commit 9045538339885234729d50324c5dc6036cfc568f (#194698).
"guichan: remove formula"

This reverts commit 54732ff7981bbf6e2b055370967cc5027eabefee (#151272).
"guichan: disable"

This reverts commit b3b42951b6368306b9e9748581000f1aeaa1f3a2 (#122334).
"guichan: deprecate"