act_runner 0.2.11 (new formula)
This commit adds a formula for Gitea's `act_runner`. Although `act_runner` is
available as a Docker container, this method of distribution doesn't allow the
action runner to run macOS-specific workflows (building an iOS app, etc.) due
to the containerization of the runner. Distributing the runner via Homebrew
allows workflow execution on the host with more convenient installation and
updates.