sox: upgrade SoX to SoX_NG 14.6.0.1
SoX was last updated 2015-02-22. SoX_ng project has resumed development with new releases that are
drop-in backwards-compatible. From the project homepage: https://codeberg.org/sox_ng/sox_ng

The SoX_ng project imports, compares and refines bug fixes and new work from
the 50-odd software distributions that package SoX and from the plethora of
forks on github and elsewhere and makes regular releases with a six-monthly
cadence for each of the micro (bug fixes) and minor (new features) releases.
A major release (non-backwards-compatible changes) is not planned.

Verified these changes to the formula:

- license: GPL 2.0 as per https://codeberg.org/sox_ng/sox_ng/src/branch/main/COPYING
- builds cleanly without -Wno-incompatible-function-pointer-types
- patch beginning "Fix -flat_namespace being used on Big Sur and later" is
  already integrated in SoX_ng.
- patch beginning "Applies Eric Wong's patch to fix device name length in MacOS"
  does not appear to be required.  coreaudio.c has been significantly changed,
  a different API is used to get device names, and device names appear to be
  retrieved correctly with sox -V -t coreaudio null -n.
- added dependency on wavpack and verfied SoX_ng builds with wavpack support.