media-control: Test with the test command
The test command not only tests that the tool works, but also that the MediaRemote framework, which is used under the hood to detect media playback on the device, is still accessible. The test command runs "get" internally, so the old way of testing is still integrated. The "test" command is the better alternative because it will actually fail, when the MediaRemote framework is inaccessible.