mcp-server-kubernetes: better testing
I think that the current tests actually do not work properly.
A MCP server process is a daemon that terminate on SIGTERM so i think
it currently fails but with somehow the searched strings in the output.

The cause of the failure is likely the `notifications/initialized`
client anwser that is missing meanwhile it is required by MCP version
`2025-03-26`.

I have catched this error while trying to write other MCP formulae
(PRs coming soon).

Proposed fix:
- send the missing message
- send the `TERM` signal (otherwise process keep running)
- parse the output to be sure to have right answer