macvim: sign and use correct runtime entitlements
By default, Homebrew was building MacVim with a default ad-hoc signature
with get-task-allow enabled. This could have security implications as it
allows other apps to hook into the app and inspect/interfere with its
operation. MacVim has added a new Makefile target specifically for
signing the app with the correct runtime hardening and entitlements with
an ad-hoc signature. Use it here so Homebrew's version of MacVim has the
correct runtime entitlements.