Add workaround for OpenSSL 3.5 API changes
This pulls in a PR from the lastpass-cli repository that primarily fixes the ability to write notes.

Without it, the cli tool breaks frequently with an error about being unable to base64 decode the response - this is because lastpass-cli without the change blindly passes all input from the attachment_key field to OpenSSL.

Older versions of OpenSSL silently dealt with this, newer ones don't so the patch referenced is a 1 line change to make sure that if the string that would be passed to OpenSSL is "skipped", the string is not passed to OpenSSL so the same code path is followed under all versions.