mariadb@11.8: Use accurate end-of-life dates
As MariaDB is built from sources on Homebrew, the actual end-of-life dates
announced on Homebrew should reflect the date MariaDB is committed to
publish updates as source code, which is further out than what they commit
to regarding binaries.

With the updates in this commit and the two preceding it all current LTS
versions of MariaDB have correct EOL dates on Homebrew and the Formulaes
helpful comments to make the upstream policy clear for maintainers.